SCB-1391 Polish code with suggestion of Willem Jiang
