+++ omega/omega-transaction/src/main/java/org/apache/servicecomb/pack/omega/transaction/TransactionAspect.java
@@ -30,11 +30,9 @@ import org.aspectj.lang.reflect.MethodSignature;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.apache.servicecomb.pack.omega.transaction.TransactionContextUtils.extractTransactionContext;
-import static org.apache.servicecomb.pack.omega.transaction.TransactionContextUtils.populateOmegaContext;
-
 @Aspect
-public class TransactionAspect {
+public class TransactionAspect extends TransactionContextHelper {
+
   private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
   private final OmegaContext context;
@@ -53,7 +51,7 @@ public class TransactionAspect {
     // just check if we need to setup the transaction context information first
     TransactionContext transactionContext = extractTransactionContext(joinPoint.getArgs());
     if (transactionContext != null) {
-      populateOmegaContext(context, transactionContext, LOG);
+      populateOmegaContext(context, transactionContext);
     }
 
     String localTxId = context.localTxId();
@@ -70,5 +68,9 @@ public class TransactionAspect {
     }
   }
 
+  @Override
+  protected Logger getLogger() {
+    return LOG;
+  }
 
 }
