+++ omega/omega-transaction/src/main/java/org/apache/servicecomb/pack/omega/transaction/tcc/TccParticipatorAspect.java
@@ -19,9 +19,11 @@ package org.apache.servicecomb.pack.omega.transaction.tcc;
 
 import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Method;
+
 import org.apache.servicecomb.pack.common.TransactionStatus;
 import org.apache.servicecomb.pack.omega.context.OmegaContext;
 import org.apache.servicecomb.pack.omega.context.TransactionContext;
+import org.apache.servicecomb.pack.omega.transaction.TransactionContextHelper;
 import org.apache.servicecomb.pack.omega.transaction.annotations.Participate;
 import org.apache.servicecomb.pack.omega.transaction.tcc.events.ParticipationEndedEvent;
 import org.apache.servicecomb.pack.omega.transaction.tcc.events.ParticipationStartedEvent;
@@ -32,11 +34,10 @@ import org.aspectj.lang.reflect.MethodSignature;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.apache.servicecomb.pack.omega.transaction.TransactionContextUtils.extractTransactionContext;
-import static org.apache.servicecomb.pack.omega.transaction.TransactionContextUtils.populateOmegaContext;
 
 @Aspect
-public class TccParticipatorAspect {
+public class TccParticipatorAspect extends TransactionContextHelper {
+
   private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
   private final OmegaContext context;
@@ -58,13 +59,13 @@ public class TccParticipatorAspect {
     Method method = ((MethodSignature) joinPoint.getSignature()).getMethod();
     TransactionContext transactionContext = extractTransactionContext(joinPoint.getArgs());
     if (transactionContext != null) {
-      populateOmegaContext(context, transactionContext, LOG);
+      populateOmegaContext(context, transactionContext);
     }
 
     String localTxId = context.localTxId();
     String cancelMethod = callbackMethodSignature(joinPoint, participate.cancelMethod(), method);
     String confirmMethod = callbackMethodSignature(joinPoint, participate.confirmMethod(), method);
-    
+
     context.newLocalTxId();
     LOG.debug("Updated context {} for participate method {} ", context, method.toString());
 
@@ -98,5 +99,8 @@ public class TccParticipatorAspect {
         .toString();
   }
 
-
+  @Override
+  protected Logger getLogger() {
+    return LOG;
+  }
 }
